Make reusable useKeyDown
